feat: user수 관리 포인트를 Users에서 관리하도록 수정
