+++ src/main/java/nextstep/ladder/domain/line/Line.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain.line;
 
+import nextstep.ladder.domain.user.Users;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -7,13 +9,10 @@ public class Line {
 
     private List<Boolean> points = new ArrayList<>();
 
-    public Line(int userCount, LineCreateStrategy lineCreateStrategy) {
-        if (userCount <= 1) {
-            throw new IllegalArgumentException("유저는 2명 이상이여야 합니다.");
-        }
-
+    public Line(Users users, LineCreateStrategy lineCreateStrategy) {
         points.add(lineCreateStrategy.createFirstPoint());
-        for (int i = 1; i < userCount; i++) {
+
+        for (int i = 1; i < users.getSize(); i++) {
             boolean nextPoint = lineCreateStrategy.createNextPoint(points.get(i - 1));
             points.add(nextPoint);
         }
