+++ src/main/java/nextstep/ladder/domain/user/Users.java
@@ -9,5 +9,13 @@ public class Users {
 
     public Users(String names) {
         this.users = Stream.of(names.split(",")).map(User::new).collect(Collectors.toList());
+
+        if (this.users.size() <= 1) {
+            throw new IllegalArgumentException("유저는 2명 이상이여야 합니다.");
+        }
+    }
+
+    public int getSize() {
+        return users.size();
     }
 }
