+++ src/test/java/nextstep/ladder/domain/line/LinesTest.java
@@ -1,5 +1,6 @@
 package nextstep.ladder.domain.line;
 
+import nextstep.ladder.domain.user.UsersTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -23,13 +24,13 @@ class LinesTest {
     @DisplayName("Lines 인스턴스 만들기")
     @Test
     public void testConstructor() {
-        assertDoesNotThrow(() -> new Lines(3, 2, lineCreateStrategyStub));
+        assertDoesNotThrow(() -> new Lines(UsersTest.USERS1, 2, lineCreateStrategyStub));
     }
 
     @DisplayName("높이가 0 이하이면 예외를 반환한다.")
     @Test
     public void testConstructor_throwException() {
-        assertThatThrownBy(() -> new Lines(3, 0, lineCreateStrategyStub))
+        assertThatThrownBy(() -> new Lines(UsersTest.USERS1, 0, lineCreateStrategyStub))
             .isInstanceOf(IllegalArgumentException.class)
             .hasMessageContaining("높이는 1 이상이여야 합니다.");
     }
